+++ public/example/exp-month-dailynotes.js
@@ -1,0 +1,9 @@
+//!js
+const monthStart = Query.utils.thisMonth(false);
+const today = Query.utils.today(false);
+const dailyNotes = await Query.dailynote({
+    after: monthStart,
+    before: today,
+    limit: 32
+});
+return dailyNotes.pick('id');